+++ lua/core/menu/tape.lua
@@ -1,0 +1,194 @@
+local TAPE_MODE_PLAY = 1
+local TAPE_MODE_REC = 2
+
+local TAPE_PLAY_LOAD = 1
+local TAPE_PLAY_PLAY = 2
+local TAPE_PLAY_STOP = 3
+local TAPE_PLAY_PAUSE = 4
+local TAPE_REC_ARM = 1
+local TAPE_REC_START = 2
+local TAPE_REC_STOP = 3
+
+
+local m = {}
+m.mode = TAPE_MODE_PLAY
+m.play = {}
+m.play.sel = TAPE_PLAY_LOAD
+m.play.status = TAPE_PLAY_STOP
+m.play.file = nil
+m.play.pos_tick = 0
+m.rec = {}
+m.rec.file = nil
+m.rec.sel = TAPE_REC_ARM
+m.rec.pos_tick = 0
+m.diskfree = 0
+
+local DISK_RESERVE = 250
+local function tape_diskfree()
+  if norns.disk then
+    m.diskfree = math.floor((norns.disk - DISK_RESERVE) / .192) -- seconds of 48k/16bit stereo disk free with reserve
+  end
+end
+
+
+m.key = function(n,z)
+  if n==2 and z==1 then
+    m.mode = (m.mode==1) and 2 or 1
+    menu.redraw()
+  elseif n==3 and z==1 then
+    if m.mode == TAPE_MODE_PLAY then
+      if m.play.sel == TAPE_PLAY_LOAD then
+        local playfile_callback = function(path)
+          if path ~= "cancel" then
+            audio.tape_play_open(path)
+            m.play.file = path:match("[^/]*$")
+            m.play.status = TAPE_PLAY_PAUSE
+            m.play.sel = TAPE_PLAY_PLAY
+            local _, samples, rate = sound_file_inspect(path)
+            m.play.length = math.floor(samples / rate)
+            m.play.length_text = util.s_to_hms(m.play.length)
+            m.play.pos_tick = 0
+            tape_play_counter.time = 0.25
+            tape_play_counter.event = function()
+              m.play.pos_tick = m.play.pos_tick + 0.25
+              if m.play.pos_tick > m.play.length
+                  and m.play.status == TAPE_PLAY_PLAY then
+                print("tape is over!")
+                audio.tape_play_stop()
+                tape_play_counter:stop()
+                m.play.file = nil
+                m.play.stats = TAPE_PLAY_STOP
+                m.play.sel = TAPE_PLAY_LOAD
+              end
+              if menu.mode == true and menu.page == pTAPE then
+                menu.redraw()
+              end
+            end
+          else
+            m.play.file = nil
+          end
+          menu.redraw()
+        end
+        fileselect.enter(_path.audio, playfile_callback)
+      elseif m.play.sel == TAPE_PLAY_PLAY then
+        tape_play_counter:start()
+        audio.tape_play_start()
+        m.play.status = m.play.sel
+        m.play.sel = TAPE_PLAY_STOP
+        menu.redraw()
+      elseif m.play.sel == TAPE_PLAY_STOP then
+        audio.tape_play_stop()
+        tape_play_counter:stop()
+        m.play.file = nil
+        m.play.status = m.play.sel
+        m.play.sel = TAPE_PLAY_LOAD
+        menu.redraw()
+      end
+    else -- REC CONTROLS
+      if m.rec.sel == TAPE_REC_ARM then
+        tape_diskfree()
+        m.rec.file = string.format("%04d",norns.state.tape) .. ".wav"
+        audio.tape_record_open(_path.audio.."tape/"..m.rec.file)
+        m.rec.sel = TAPE_REC_START
+        m.rec.pos_tick = 0
+        tape_rec_counter.time = 0.25
+        tape_rec_counter.event = function()
+          m.rec.pos_tick = m.rec.pos_tick + 0.25
+          if m.rec.pos_tick > m.diskfree then
+            print("out of space!")
+            audio.tape_record_stop()
+            norns.state.tape = norns.state.tape + 1
+            tape_rec_counter:stop()
+            m.rec.sel = TAPE_REC_ARM
+          end
+          if menu.mode == true and menu.page == pTAPE then
+            menu.redraw()
+          end
+        end
+      elseif m.rec.sel == TAPE_REC_START then
+        tape_rec_counter:start()
+        audio.tape_record_start()
+        m.rec.sel = TAPE_REC_STOP
+      elseif m.rec.sel == TAPE_REC_STOP then
+        tape_rec_counter:stop()
+        norns.state.tape = norns.state.tape + 1
+        audio.tape_record_stop()
+        m.rec.sel = TAPE_REC_ARM
+        tape_diskfree()
+      end
+      menu.redraw()
+    end
+  end
+end
+
+m.enc = norns.none
+
+m.redraw = function()
+  screen.clear()
+
+  menu.draw_panel()
+
+  screen.move(128,10)
+	screen.level(m.mode==TAPE_MODE_PLAY and 15 or 1)
+	screen.text_right("PLAY")
+  screen.level(2)
+  screen.rect(0.5,13.5,127,2)
+  screen.stroke()
+
+  if m.play.file then
+    screen.level(2)
+    screen.move(0,10)
+    screen.text(m.play.file)
+    screen.move(0,24)
+    screen.text(util.s_to_hms(math.floor(m.play.pos_tick)))
+    screen.move(128,24)
+    screen.text_right(m.play.length_text)
+    screen.level(15)
+    screen.move((m.play.pos_tick / m.play.length * 128),13.5)
+    screen.line_rel(0,2)
+    screen.stroke()
+    if m.mode==TAPE_MODE_PLAY then
+      screen.level(15)
+      screen.move(64,24)
+      if m.play.sel == TAPE_PLAY_PLAY then screen.text_center("START")
+      elseif m.play.sel == TAPE_PLAY_STOP then screen.text_center("STOP") end
+    end
+  end
+
+  screen.move(128,48)
+  screen.level(m.mode==TAPE_MODE_REC and 15 or 1)
+  screen.text_right("REC")
+  screen.level(2)
+  screen.rect(0.5,51.5,127,2)
+  screen.stroke()
+  if m.mode==TAPE_MODE_REC then
+    screen.level(15)
+    screen.move(64,62)
+    if m.rec.sel == TAPE_REC_START then screen.text_center("START")
+    elseif m.rec.sel == TAPE_REC_STOP then screen.text_center("STOP") end
+  end
+  if m.rec.sel ~= TAPE_REC_ARM then
+    screen.level(1)
+    screen.move(0,48)
+    screen.text(string.format("%04d",norns.state.tape))
+    screen.level(2)
+    screen.move(0,62)
+    screen.text(util.s_to_hms(math.floor(m.rec.pos_tick)))
+  end
+  screen.level(2)
+  screen.move(127,62)
+  screen.text_right(util.s_to_hms(m.diskfree))
+  screen.level(15)
+  screen.move((m.rec.pos_tick / m.diskfree * 128),51.5)
+  screen.line_rel(0,2)
+  screen.stroke()
+
+  screen.update()
+end
+
+m.init = function()
+  tape_diskfree()
+end
+m.deinit = norns.none
+
+return m
