+++ lua/core/menu/wifi.lua
@@ -1,0 +1,123 @@
+local m = {}
+m.pos = 0
+m.list = {"off","hotspot","connect", "add", "del"}
+m.len = #m.list
+m.selected = 1
+m.try = ""
+m.countdown = -1
+
+m.connect = function(x)
+  if x ~= "cancel" then
+    wifi.on(x)
+  end
+  menu.redraw()
+end
+
+m.add = function(x)
+  m.try = x
+  if x ~= "cancel" then
+    textentry.enter(m.passdone, "", "enter password:")
+  end
+  menu.redraw()
+end
+
+m.del = function(x)
+  if x ~= "cancel" then
+    wifi.delete(x)
+  end
+  menu.redraw()
+end
+
+m.passdone = function(txt)
+  if txt ~= nil then
+    print("adding " .. m.try .. txt)
+    wifi.add(m.try, txt)
+  end
+  menu.redraw()
+end
+
+
+m.key = function(n,z)
+  if n==2 and z==1 then
+    menu.set_page("SYSTEM")
+  elseif n==3 and z==1 then
+    if m.pos == 0 then
+      wifi.off()
+    elseif m.pos == 1 then
+      wifi.hotspot()
+    elseif m.pos == 2 then
+      wifi.update()
+      listselect.enter(wifi.conn_list, m.connect)
+    elseif m.pos == 3 then
+      wifi.update()
+      listselect.enter(m.ssid_list, m.add)
+    elseif m.pos == 4 then
+      wifi.update()
+      listselect.enter(wifi.conn_list, m.del)
+    end
+  end
+end
+
+m.enc = function(n,delta)
+  if n==2 then
+    m.pos = m.pos + delta
+    if m.pos > m.len - 1 then m.pos = m.len - 1
+    elseif m.pos < 0 then m.pos = 0 end
+    menu.redraw()
+  elseif n==3 and m.pos == 2 then
+    m.selected = util.clamp(1,m.selected+delta,wifi.conn_count)
+    menu.redraw()
+  end
+end
+
+m.redraw = function()
+  screen.clear()
+  screen.level(4)
+
+  screen.move(0,10)
+  screen.text("STATUS: " .. wifi.status)
+  screen.move(0,20)
+  screen.text("NETWORK: " .. wifi.connection_name)
+  screen.move(0,30)
+  screen.text("IP: " .. wifi.ip)
+  if wifi.ip and wifi.connection then
+    if wifi.connection:is_wireless() then
+      screen.move(0,40)
+      screen.text("SIGNAL: " .. wifi.signal .. "dBm")
+    end
+  end
+
+  local xp = {0,20,58,94,114}
+  for i=1,m.len do
+    screen.move(xp[i],60)
+    local line = m.list[i]
+    if(i==m.pos+1) then
+      screen.level(15)
+    else
+      screen.level(4)
+    end
+    screen.text(string.upper(line))
+  end
+
+  screen.update()
+end
+
+m.init = function()
+  wifi.ensure_radio_is_on()
+  m.ssid_list = wifi.ssids() or {}
+  wifi.update()
+  m.selected = 1
+  menu.timer.time = 3
+  menu.timer.count = -1
+  menu.timer.event = function()
+    wifi.update()
+    menu.redraw()
+  end
+  menu.timer:start()
+end
+
+m.deinit = function()
+  menu.timer:stop()
+end
+
+return m
